% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_interaction_helper_functions.R
\name{bootstrap_T_obs}
\alias{bootstrap_T_obs}
\title{Internal: Bootstrap Interaction Statistic (Adaptive or Fixed)}
\usage{
bootstrap_T_obs(
  X,
  Y,
  MX,
  MY,
  g_col,
  a_col,
  g1,
  g2,
  B = NULL,
  seed = NULL,
  min_iter = 500,
  max_iter = 10000,
  tol = 0.001,
  batch_size = 100,
  check_convergence = TRUE
)
}
\description{
Computes bootstrapped standard errors and 95% confidence intervals
for observed interaction statistics using nonparametric sampling within ancestries.
}
\keyword{internal}
