% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/split_stratified_ancestry_sets_helper_functions.R
\name{check_strata_feasibility}
\alias{check_strata_feasibility}
\title{Check Stratification Feasibility}
\usage{
check_strata_feasibility(strata_X, strata_Y)
}
\arguments{
\item{strata_X}{A factor for overrepresented group strata (e.g., EUR).}

\item{strata_Y}{A factor for underrepresented group strata (e.g., AFR).}
}
\value{
A list with `usable`, `missing`, and `insufficient` strata names.
}
\description{
Identifies usable, missing, or insufficient strata based on availability in each group.
}
\keyword{internal}
