% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_diff_interaction_helper_functions.R
\name{compute_CI_bounds}
\alias{compute_CI_bounds}
\title{Confidence Interval Bounds}
\usage{
compute_CI_bounds(T_boot)
}
\arguments{
\item{T_boot}{A matrix of bootstrapped test statistics.}
}
\value{
A matrix of lower and upper confidence bounds.
}
\description{
Computes 2.5% and 97.5% quantiles for each feature.
}
\keyword{internal}
