% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_aggregating_sets_helper_functions.R
\name{compute_empirical_p}
\alias{compute_empirical_p}
\title{Empirical P-Values}
\usage{
compute_empirical_p(
  T_obs,
  T_null,
  alternative = c("two.sided", "greater", "less")
)
}
\arguments{
\item{T_obs}{A numeric vector of observed test statistics.}

\item{T_null}{A numeric vector of null distribution test statistics.}
}
\value{
A numeric vector of p-values.
}
\description{
Computes two-sided empirical p-values.
}
\keyword{internal}
