% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_interaction_helper_functions.R
\name{compute_empirical_pvalues}
\alias{compute_empirical_pvalues}
\title{Empirical P-Values}
\usage{
compute_empirical_pvalues(T_boot, T_obs)
}
\arguments{
\item{T_boot}{A matrix of bootstrapped test statistics.}

\item{T_obs}{A numeric vector of observed test statistics.}
}
\value{
A numeric vector of p-values.
}
\description{
Computes two-sided empirical p-values.
}
\keyword{internal}
