% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_diff_interaction_helper_functions.R
\name{compute_T_stat}
\alias{compute_T_stat}
\title{Compute Test Statistic for Permutation}
\usage{
compute_T_stat(XY, g_XY, p1, p2, g1, g2)
}
\arguments{
\item{XY}{A matrix of features.}

\item{g_XY}{A factor of group labels corresponding to rows of `XY`.}

\item{p1}{Indices for the first permutation group.}

\item{p2}{Indices for the second permutation group.}

\item{g1}{First group label.}

\item{g2}{Second group label.}
}
\value{
A numeric vector of test statistics or NA values.
}
\description{
Computes the interaction effect: difference of group-level differences.
Returns NA vector if any subgroup is empty.
}
\keyword{internal}
