% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nbinom_quantiles.R
\name{estimate_nbinom_quantiles}
\alias{estimate_nbinom_quantiles}
\title{Negative Binomial quantiles per gene}
\usage{
estimate_nbinom_quantiles(mu, phi, probs)
}
\arguments{
\item{mu}{Numeric vector of means (>0).}

\item{phi}{Numeric vector of dispersions (>0), same length as \code{mu}.}

\item{probs}{Probabilities in [0, 1] at which to compute quantiles.}
}
\value{
Matrix of quantiles with rows = \code{probs} and cols = genes.
}
\description{
Computes quantiles of the Negative Binomial distribution for each gene,
given mean (\code{mu}) and dispersion (\code{phi}) values.
}
