% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc_build_beta_matrix.R
\name{gdc_build_beta_matrix}
\alias{gdc_build_beta_matrix}
\title{Build GDC Beta-values count matrix}
\usage{
gdc_build_beta_matrix(file_map, file_dir, verbose = TRUE)
}
\arguments{
\item{file_map}{Data frame with FILE_ID, FILE_NAME, SAMPLE_ID.}

\item{file_dir}{Directory containing GDC subfolders.}

\item{verbose}{Print progress messages.}
}
\value{
Numeric matrix (samples × features).
}
\description{
Reads GDC Illumina 450k files and builds a numeric matrix with
samples as rows and features as columns.
}
