% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc_gene_level_count_matrix.R
\name{gdc_gene_level_count_matrix}
\alias{gdc_gene_level_count_matrix}
\title{Collapse duplicate genes with sum and remove NAs from count matrix}
\usage{
gdc_gene_level_count_matrix(matr, verbose = TRUE)
}
\arguments{
\item{matr}{Numeric matrix (samples × features).}

\item{verbose}{Logical; show summary (default TRUE).}
}
\value{
Matrix with NAs removed and duplicate genes summed.
}
\description{
Collapse duplicate genes with sum and remove NAs from count matrix
}
