% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc_omics_layer.R
\name{gdc_omics_layer}
\alias{gdc_omics_layer}
\title{Align metadata and matrix by IDs}
\usage{
gdc_omics_layer(meta, matr, id, tech, verbose = TRUE)
}
\arguments{
\item{meta}{Metadata table.}

\item{matr}{Omics matrix.}

\item{id}{ID column name.}

\item{tech}{Technology name.}

\item{verbose}{Show messages.}
}
\value{
Named list with meta and matr.
}
\description{
Reorders matrix rows to match metadata.
Checks ID sets for consistency.
}
