% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/split_stratified_ancestry_sets_helper_functions.R
\name{get_strata}
\alias{get_strata}
\title{Create Stratification Factor}
\usage{
get_strata(meta, stratify_cols)
}
\arguments{
\item{meta}{A data.frame containing metadata.}

\item{stratify_cols}{A character vector of column names to stratify by.}
}
\value{
A factor indicating stratification strata.
}
\description{
Combines metadata columns into a factor for stratified sampling.
}
\keyword{internal}
