% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_prediction.R
\name{logistic_regression_prediction}
\alias{logistic_regression_prediction}
\title{Logistic regression prediction}
\usage{
logistic_regression_prediction(
  R,
  X,
  Y,
  MR,
  MX,
  MY,
  g_col,
  a_col,
  n_folds,
  n_models,
  maxit = NULL,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{R}{Expression matrix for refeernce ancestry. Rows = samples, columns = genes.}

\item{X}{Expression matrix for ancestry X. Rows = samples, columns = genes.}

\item{Y}{Expression matrix for ancestry Y. Rows = samples, columns = genes.}

\item{MR}{Metadata for R. Must include group and ancestry columns.}

\item{MX}{Metadata for X. Must include group and ancestry columns.}

\item{MY}{Metadata for Y. Must include group and ancestry columns.}

\item{g_col}{Name of the column indicating group (factor 1).}

\item{a_col}{Name of the column indicating ancestry (factor 2).}

\item{n_folds}{CV folds}

\item{n_models}{Grid size}

\item{maxit}{Max iter}

\item{seed}{RNG seed}

\item{verbose}{Print msgs}
}
\value{
A list with:
\describe{
  \item{summary_stats}{Predictions + labels}
  \item{feature_stats}{Non-zero glmnet coefs}
}
}
\description{
Fits elastic-net logistic models on R, tests on X and Y,
tunes hyperparameters, and returns
predictions and model coefficients.
}
