% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_interaction_helper_functions.R
\name{permute_indices}
\alias{permute_indices}
\title{Random Permutation of Indices}
\usage{
permute_indices(n_total, n1)
}
\arguments{
\item{n_total}{Total number of samples.}

\item{n1}{Size of the first group.}
}
\value{
A list with two index vectors: `p1` and `p2`.
}
\description{
Returns a random split of indices into two groups.
}
\keyword{internal}
