% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_interaction_effect_demographics.R
\name{plot_sim_interaction_effect_demographics}
\alias{plot_sim_interaction_effect_demographics}
\title{Plot feature-level summary using sim_imbalanced_groups summary output}
\usage{
plot_sim_interaction_effect_demographics(
  summary_list,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  fill_name = NULL
)
}
\arguments{
\item{summary_list}{The `summary` list from `sim_imbalanced_groups`.}

\item{title}{Optional plot title.}
}
\value{
A ggplot object.
}
\description{
Plot feature-level summary using sim_imbalanced_groups summary output
}
