% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trade_univariate.R
\name{run_trade_univariate}
\alias{run_trade_univariate}
\title{Run univariate TRADE analysis for one contrast}
\usage{
run_trade_univariate(stats, coef_id = NULL)
}
\arguments{
\item{stats}{Data frame of stats with required cols.}

\item{coef_id}{Contrast ID to analyze (matches coef_id col).}
}
\value{
TRADE result object (list).
}
\description{
Run univariate TRADE analysis for one contrast
}
