% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/split_stratified_ancestry_sets_helper_functions.R
\name{stratified_sample_ids}
\alias{stratified_sample_ids}
\title{Stratified Sampling by Sample IDs}
\usage{
stratified_sample_ids(strata, target_counts, ids)
}
\arguments{
\item{strata}{A factor indicating strata for each sample (must match `ids` length).}

\item{target_counts}{A named vector of how many samples to draw per stratum.}

\item{ids}{A character or numeric vector of sample IDs, same length as `strata`.}
}
\value{
A character or numeric vector of sampled IDs.
}
\description{
Performs stratified sampling from a vector of sample IDs grouped by strata.
}
\keyword{internal}
