% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/split_stratified_ancestry_sets_helper_functions.R
\name{stratified_sample_indices}
\alias{stratified_sample_indices}
\title{Stratified Sampling Indices}
\usage{
stratified_sample_indices(strata, target_counts)
}
\arguments{
\item{strata}{A factor representing sample strata (from overrepresented group).}

\item{target_counts}{A named vector of counts to sample per stratum.}
}
\value{
An integer vector of row indices to sample.
}
\description{
Performs stratified random sampling from a vector of strata.
}
\keyword{internal}
