% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_limma_interaction_effect_subsets.R
\name{summarize_limma_interaction_effect_subsets}
\alias{summarize_limma_interaction_effect_subsets}
\title{Summarize subsets results by feature/coefficient with multiple aggregation methods}
\usage{
summarize_limma_interaction_effect_subsets(
  stats,
  method = c("mean", "cct", "fisher", "bonferroni"),
  by = NULL
)
}
\arguments{
\item{stats}{A data frame with columns: `feature`, `T_obs`, `p_value`, `p_adj`, `ave_expr`.}

\item{method}{Aggregation method to highlight/return separately. Options: "mean", "cct", "fisher", "bonferroni".}

\item{by}{Additional grouping variables besides `feature` (character vector).}
}
\description{
Summarize subsets results by feature/coefficient with multiple aggregation methods
}
