% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_diff_interaction_helper_functions.R
\name{validate_ancestry}
\alias{validate_ancestry}
\title{Validate Ancestry Constraints}
\usage{
validate_ancestry(a_X, a_Y)
}
\arguments{
\item{a_X}{An ancestry label for dataset X.}

\item{a_Y}{An ancestry label for dataset Y.}
}
\description{
Ensures each dataset has exactly one ancestry level.
}
\keyword{internal}
