% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_interaction_helper_functions.R
\name{validate_groups}
\alias{validate_groups}
\title{Validate Group Factors}
\usage{
validate_groups(g_X, g_Y)
}
\arguments{
\item{g_X}{A factor vector representing group labels for dataset X.}

\item{g_Y}{A factor vector representing group labels for dataset Y.}
}
\description{
Ensures group labels are factors, share levels, and have exactly 2 levels.
}
\keyword{internal}
