% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc_download_manifest.R
\name{gdc_download_manifest}
\alias{gdc_download_manifest}
\title{Download data from GDC via gdc-client}
\usage{
gdc_download_manifest(
  manifest_file,
  metadata_file,
  token_file = NULL,
  download = TRUE,
  file_dir = NULL,
  client_path,
  verbose = TRUE
)
}
\arguments{
\item{manifest_file}{Path to GDC manifest txt file}

\item{metadata_file}{Path to GDC metadata json file}

\item{token_file}{Optional path to token file}

\item{download}{Wether to download (default: True)}

\item{file_dir}{Output dir, defaults to getwd()}

\item{client_path}{Path to gdc-client binary}

\item{verbose}{Print progress messages}
}
\value{
ID to file
}
\description{
Reads a manifest file and downloads all files
Uses current dir if out_dir is not specified
Supports optional token for protected data
Prints file and download counts when verbose
}
