% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stratified_sets.R
\name{plot_stratified_sets}
\alias{plot_stratified_sets}
\title{Plot stratified ancestry sets.}
\usage{
plot_stratified_sets(
  MX,
  MY,
  MR,
  x_var,
  fill_var,
  title = NULL,
  x_label = NULL,
  y_label = NULL
)
}
\arguments{
\item{MX}{A data frame of train meta.}

\item{MY}{A data frame of test meta.}

\item{MR}{A data frame of inference meta.}

\item{x_var}{Name of the variable for the x-axis (string).}

\item{fill_var}{Name of the variable for the fill color (string).}

\item{title}{Plot title (optional).}

\item{x_label}{Label for the x-axis (optional).}

\item{y_label}{Label for the y-axis (optional).}
}
\value{
A ggplot object.
}
\description{
Create a bar plot comparing stratified ancestry sets.
}
