% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stratified_sets.R
\name{plot_stratified_sets}
\alias{plot_stratified_sets}
\title{Plot Sample Size by Stratified Variable and Dataset Split}
\usage{
plot_stratified_sets(x, stratify_cols, point_size = 0.5)
}
\arguments{
\item{stratify_cols}{Character vector of metadata column names used for stratification.}

\item{point_size}{Numeric value to control fontsize.}

\item{result}{The output from `split_stratified_ancestry_sets()`.}
}
\value{
A ggplot2 object with dodged bars grouped by set, filled by strata values, and faceted by stratify column.
}
\description{
Visualizes the number of samples in train, test, and inference sets,
with bars grouped by dataset split and filled by stratum values. Each
stratification column gets its own facet for clearer interpretation.
}
