% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_subsets.R
\name{summarize_subsets}
\alias{summarize_subsets}
\title{Summarize subsets results by feature, with Cauchy-combined p-value}
\usage{
summarize_subsets(data, iter_col)
}
\arguments{
\item{data}{A data frame with columns: `feature`, `T_obs`, `p_value`, `p_adj`, `ave_expr`, `iteration`.}

\item{iter_col}{Name of the iteration column (string).}
}
\description{
Summarize subsets results by feature, with Cauchy-combined p-value
}
